requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=2.1.0
feedparser>=6.0.0
newspaper3k>=0.2.8
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
